pulse: set SHM size to 8MiB
The default is 64MiB and with the value set this high, pulseaudio
eventually runs out of space to create additional shm connections,
*probably* due to address space limitations on 32 bit systems.

This is why the devices ringtone would sometimes not sound, or why
openmediaplayer wouldn't produce sound after a phone call.

Fix it by setting the value to 8MiB. This has implications for N900
audio calls, but we will address that in libcmtspeechdata.